Ability to print to stdout
Added in the ability to print to stdout. status() now takes an int, 0
meaning that it will print to the status bar, anything else making it
print out to stdout. May add in more, though with text output it can
pretty easily be piped to other bars.

get_netusage has a sleep(1) in its function, so having it in the
status() func was making it take 2 seconds to update instead of 1. Fixed
that by changing to sleep(0) in status().

Signed-off-by: William Giokas <1007380@gmail.com>